vapoursynth: add withPlugins interface
Since nixpkgs does not currently have an attribute set with plugins, the
usage differes from python-style withPackages as it accepts a list
instead of a function as a parameter.

Using plugins in other programs that require vapoursynth (like `vsedit`)
can be achieved by adding it to the plugin list (e.g.
`vapoursynth.withPlugins [ vapoursynth-editor … ]`) or by passing the
`VAPOURSYNTH_PLUGIN_PATH` variable.